requirements.txt
numpy==1.26.3
langgraph==0.3.20
langchain>=0.1.1
openai==1.0.0
flask==2.3.2
python-dotenv==1.0.0
asyncio==3.4.3